- return exact index when getting an error.
- returns `ResultPos` for `Attributes` iterator

speedup:
```
before
test bench_quick_xml            ... bench:     692,552 ns/iter (+/- 8,580)
test bench_quick_xml_namespaced ... bench:     975,165 ns/iter (+/- 8,571)

after
test bench_quick_xml            ... bench:     549,643 ns/iter (+/- 11,789)
test bench_quick_xml_namespaced ... bench:     830,328 ns/iter (+/- 17,818)
```